Quick 'n ditry tweak to display commit messages on multiple lines
It's customary to have Git commit messages consist of a short, one-line
description (acting as the summary) of the change on the first line with
any further descriptions handled in paragraphs below. For a further
discussion on this point, see http://www.tpope.net/node/106

Within the commit view, rendering @commit.message appeared to strip the
newlines provided by Grit::Commit#message, so I figured splitting the
message string on the newline character and iterating over the array,
inserting a <br/> element, would do the trick.